Commit 7: logs y responseEntity. Los métodos DELETE siguen sin funcionar

+++ src/main/java/com/conanthelibrarian/librarymanagementsystem/service/UserService.java
@@ -8,6 +8,7 @@ import lombok.AllArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 
+import javax.swing.text.html.Option;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -20,23 +21,22 @@ public class UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
 
-    //todo LOGS
-
     public List<UserDTO> findUser(){
         return userRepository.findAll().stream()
                 .map(userMapper::userToUserDTO)
                 .collect(Collectors.toList());
     }
 
+    public Optional<UserDTO> findUserById(Integer id){
+        return userRepository.findById(id).stream()
+                .map(userMapper::userToUserDTO)
+                .findAny();
+    }
+
     public void newUser (UserDTO userDTO){
         User user = userMapper.userDTOToUser(userDTO);
         userRepository.save(user);
-    }
-
-    public List<UserDTO> findUserById(Integer id){
-        return userRepository.findById(id).stream()
-                .map(userMapper::userToUserDTO)
-                .collect(Collectors.toList());
+        log.info("User saved with name: {}", userDTO.getName());
     }
 
     public UserDTO modifyUser(Integer id, UserDTO userDTO){
@@ -44,12 +44,27 @@ public class UserService {
         if(userOptional.isPresent()){
             User user = userMapper.userDTOToUser(userDTO);
             userRepository.save(user);
+            log.info("User modified with name: {}", userDTO.getName());
         }
         return userDTO;
     }
 
     public void deleteUserById(Integer id) {
-        userRepository.deleteById(id);
+        if(id != null){
+            UserDTO userDTO = UserDTO.builder().build();
+            userMapper.userDTOToUser(userDTO);
+            userRepository.deleteById(id);
+            log.info("User deleted with id: {}", userDTO.getUserId());
+        }
+        /*Error:
+        Error when deleting user: could not execute statement
+        [Cannot delete or update a parent row: a foreign key constraint fails
+        (`library`.`loans`, CONSTRAINT `loans_ibfk_1` FOREIGN KEY (`user_id`)
+        REFERENCES `users` (`user_id`))]
+        [delete from users where user_id=?];
+        SQL [delete from users where user_id=?]; constraint [null]
+         */
     }
+
 
 }
